+++ CHANGELOG.md
@@ -4,6 +4,26 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
+## [0.12.1] - 2022-08-31
+
+### Changed
+
+- Improved Error handling
+- Fix wrong executable in version 0.12.0
+
+## [0.12.0] - 2022-08-31
+
+### Added
+
+- Support for Go
+
+## [0.11.0] - 2022-08-30
+
+### Added
+
+- Support for R.
+- Embedding R plots into notes.
+
 ## [0.10.0] - 2022-08-29
 
 ### Changed